Fix tests (compilation issues; remove unnecessary casts)

+++ tests/manifold/copy_material_to_manifold_id_01.cc
@@ -34,16 +34,16 @@ void print_info(Triangulation<dim,spacedim> &tria)
     {
       deallog << "cell: " << cell
               << ", material_id: "
-              << (int)cell->material_id()
+              << cell->material_id()
               << ", manifold_id: "
-              << (int)cell->manifold_id() << std::endl;
+              << cell->manifold_id() << std::endl;
 
       for (unsigned int f=0; f<GeometryInfo<dim>::faces_per_cell; ++f)
         deallog << "face: " << cell->face(f)
                 << ", boundary_id: "
-                << (int)cell->face(f)->boundary_id()
+                << cell->face(f)->boundary_id()
                 << ", manifold_id: "
-                << (int)cell->face(f)->manifold_id() << std::endl;
+                << cell->face(f)->manifold_id() << std::endl;
     }
 }
 
@@ -82,4 +82,3 @@ int main ()
 
   return 0;
 }
-
